Hoist ArrayCallKind to a separate file
