+++ include/swift/SILOptimizer/Analysis/ArraySemantic.h
@@ -14,39 +14,11 @@
 #define SWIFT_SILOPTIMIZER_ANALYSIS_ARRAYSEMANTIC_H
 
 #include "swift/SIL/SILInstruction.h"
+#include "swift/SILOptimizer/Analysis/ArrayCallKind.h"
 
 namespace swift {
 
 class DominanceInfo;
-
-/// The kind of array operation identified by looking at the semantics attribute
-/// of the called function.
-enum class ArrayCallKind {
-  kNone = 0,
-  kArrayPropsIsNativeTypeChecked,
-  kCheckSubscript,
-  kCheckIndex,
-  kGetCount,
-  kGetCapacity,
-  kGetElement,
-  kGetElementAddress,
-  kMakeMutable,
-  kEndMutation,
-  kMutateUnknown,
-  kReserveCapacityForAppend,
-  kWithUnsafeMutableBufferPointer,
-  kAppendContentsOf,
-  kAppendElement,
-  // The following two semantic function kinds return the result @owned
-  // instead of operating on self passed as parameter. If you are adding
-  // a function, and it has a self parameter, make sure that it is defined
-  // before this comment.
-  kArrayInit,
-  kArrayInitEmpty,
-  kArrayUninitialized,
-  kArrayUninitializedIntrinsic,
-  kArrayFinalizeIntrinsic
-};
 
 /// Return true is the given function is an array semantics call.
 ArrayCallKind getArraySemanticsKind(SILFunction *f);
